+++ CHANGES.md
@@ -5,6 +5,8 @@ Changes in 3.9.1
 ----------------
 - Fix UpdateAttendanceDaily() call when Updating in Administration.php
 - Fix "You are not currently in a marking period" error when recalculating daily attendance in UpdateAttendanceDaily.fnc.php
+- Add Daily Comment column in TakeAttendance.php, sponsored by Asian Hope
+- Moving from github.com to gitlab.com, program wide
 
 Changes in 3.9
 --------------